pyproject.toml
[project]
name = "action-python-security-auditing-tests"
version = "0.1.0"
requires-python = ">=3.12"

[dependency-groups]
dev = [
    "pytest>=8",
    "pyyaml>=6",
    "pre-commit>=4",
]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
# Adds .github/scripts to sys.path so tests can `import validate_results`
pythonpath = [".github/scripts"]

# ---------------------------------------------------------------------------
# ruff — linting + formatting for .github/scripts/ and tests/ only
# The 01-14 directories are intentionally insecure test fixtures and must
# never be linted or reformatted.
# ---------------------------------------------------------------------------

[tool.ruff]
target-version = "py312"
line-length = 99
extend-exclude = [
    "01-requirements-flat",
    "02-requirements-src-bandit",
    "03-requirements-multi-both",
    "04-uv-flat",
    "05-uv-src-vuln",
    "06-uv-multi-bandit",
    "07-poetry-flat",
    "08-poetry-src-both",
    "09-pipenv-flat",
    "10-pipenv-multi-bandit",
    "11-requirements-root",
    "12-uv-flat-bandit-only",
    "13-requirements-unfixable",
    "14-uv-low-threshold",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "S",   # flake8-bandit (security)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # assert statements are expected in pytest tests
